Extract test modules to individual files to support cluster test execution
