+++ test/commands/support/identity/identity_aggregate_router.ex
@@ -1,0 +1,13 @@
+defmodule Commanded.Commands.IdentityAggregateRouter do
+  @moduledoc false
+  use Commanded.Commands.Router
+
+  alias Commanded.Commands.IdentityAggregate
+  alias Commanded.Commands.IdentityAggregate.IdentityCommand
+
+  identify IdentityAggregate,
+    by: :uuid,
+    prefix: "prefix-"
+
+  dispatch IdentityCommand, to: IdentityAggregate
+end
