+++ test/test_helper.exs
@@ -1,3 +1,1 @@
 ExUnit.start()
-
-{:ok, _} = Commanded.Helpers.CommandAuditMiddleware.start_link()
